refactor: move api to index
